add common logic to shared-actions mixin
