+++ app/index/controller.js
@@ -1,35 +1,11 @@
 import Ember from 'ember';
 import mapBboxController from 'mobility-playground/mixins/map-bbox-controller';
 import setTextboxClosed from 'mobility-playground/mixins/set-textbox-closed';
-import removePin from 'mobility-playground/mixins/shared-actions';
+import sharedActions from 'mobility-playground/mixins/shared-actions';
 
-export default Ember.Controller.extend(mapBboxController, setTextboxClosed, removePin, {
+export default Ember.Controller.extend(mapBboxController, setTextboxClosed, sharedActions, {
 	queryParams: ['bbox','pin'],
-	bbox: null,
-  leafletBbox: null,
-  leafletBounds: [[43.053900124340984, -89.46407318115234],[43.10875337930414, -89.32708740234375]],
-	place: null,
-  pin: null,
-  pinLocation: Ember.computed('pin', function(){
-    if (typeof(this.get('pin'))==="string"){
-      var pinArray = this.get('pin').split(',');
-      return pinArray;
-    } else {
-      return this.get('pin');
-    }
-  }),
-  attribution: '&copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors | <a href="http://www.mapzen.com">Mapzen</a> | <a href="http://www.transit.land">Transitland</a> | Imagery © <a href="https://carto.com/">CARTO</a>',
-  closeTextbox: Ember.inject.service(),
-  textboxIsClosed: Ember.computed('closeTextbox.textboxIsClosed', function(){
-    return this.get('closeTextbox').get('textboxIsClosed');
-  }),
-  currentlyLoading: Ember.inject.service(),
-	icon: L.icon({
-		iconUrl: 'assets/images/marker1.png',		
-		iconSize: (20, 20),
-    iconAnchor: [10, 24],
-	}),
-  markerUrl: 'assets/images/marker1.png',
+  
   mapCenter: [43.072963279523,-89.39234018325806],
   center: Ember.computed('pin', function(){
     if (this.get('pin')){
@@ -39,43 +15,12 @@ export default Ember.Controller.extend(mapBboxController, setTextboxClosed, remo
     }
   }),
   zoom: 14,
+
 	actions: {
 		updatebbox(e) {
 			var newbox = e.target.getBounds();
 			this.set('bbox', newbox.toBBoxString());
 		},
-  	searchRepo(term) {
-      if (Ember.isBlank(term)) { return []; }
-      const url = `https://search.mapzen.com/v1/autocomplete?api_key=mapzen-jLrDBSP&text=${term}`;      
-      return Ember.$.ajax({ url }).then(json => json.features);
-    },
-  	setPlace: function(selected){
-      this.set('pin', null);
-      var lng = selected.geometry.coordinates[0];
-      var lat = selected.geometry.coordinates[1];
-      var coordinates = [];
-      coordinates.push(lat);
-      coordinates.push(lng);
-      
-      this.set('place', selected);
-      this.set('pin', coordinates);
-      this.set('center', coordinates);
-      this.transitionToRoute('index', {queryParams: {pin: this.get('pin'), bbox: null}});
-  	},
-  	clearPlace: function(){
-  		this.set('place', null);
-  	},
-    // removePin: function(){
-    //   this.set('pin', null);
-    // },
-    dropPin: function(e){
-      var lat = e.latlng.lat;
-      var lng = e.latlng.lng;
-      var coordinates = [];
-      coordinates.push(lat);
-      coordinates.push(lng);
-      this.set('pin', coordinates);
-    },
     setIsochroneMode: function(mode){
       if (this.get('isochrone_mode') === mode){
         this.set('isochrone_mode', null);
