+++ app/operators/controller.js
@@ -1,27 +1,16 @@
 import Ember from 'ember';
 import mapBboxController from 'mobility-playground/mixins/map-bbox-controller';
 import setTextboxClosed from 'mobility-playground/mixins/set-textbox-closed';
+import sharedActions from 'mobility-playground/mixins/shared-actions';
 
-export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
+
+export default Ember.Controller.extend(mapBboxController, setTextboxClosed, sharedActions, {
 	queryParams: ['bbox', 'onestop_id','pin'],
-	pin: null,
-	pinLocation: Ember.computed('pin', function(){
-    if (typeof(this.get('pin'))==="string"){
-      var pinArray = this.get('pin').split(',');
-      return pinArray;
-    } else {
-      return this.get('pin');
-    }
-	}),
-	bbox: null,
-	leafletBbox: null,
-  leafletBounds: [[37.706911598228466, -122.54287719726562],[37.84259697150785, -122.29568481445312]],
+	
 	queryIsInactive: false,
-  currentlyLoading: Ember.inject.service(),
 	onestop_id: null,
 	selectedOperator: null,
 	hoverOperator: null,
-	place: null,
 	placeholderMessage: Ember.computed('leafletBbox', function(){
 		var total = this.model.get('meta.total');
 		if (total > 1){
@@ -39,11 +28,6 @@ export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
 			return onlyOperator;
 		}
 	}),
-	icon: L.icon({
-		iconUrl: 'assets/images/marker.png',		
-		iconSize: (20, 20),
-    iconAnchor: [10, 24],
-	}),
 	operators: Ember.computed('model', function(){
 		if (this.get('model') === null){
 			return
@@ -56,11 +40,7 @@ export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
 	}),
 	mapMoved: false,
 	mousedOver: false,
-  attribution: '&copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors | <a href="http://www.mapzen.com">Mapzen</a> | <a href="http://www.transit.land">Transitland</a> | Imagery © <a href="https://carto.com/">CARTO</a>',
-	closeTextbox: Ember.inject.service(),
-  textboxIsClosed: Ember.computed('closeTextbox.textboxIsClosed', function(){
-    return this.get('closeTextbox').get('textboxIsClosed');
-  }),
+
 	actions: {
 		setOperator(operator){
 			var onestop_id = operator.get('id');
@@ -99,37 +79,6 @@ export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
 			operator.set('operator_path_opacity', 0.5);
 			operator.set('operator_path_weight', 1.5);
 			this.set('hoverOperator', null);
-		},
-		searchRepo(term) {
-      if (Ember.isBlank(term)) { return []; }
-      const url = `https://search.mapzen.com/v1/autocomplete?api_key=mapzen-jLrDBSP&text=${term}`; 
-      return Ember.$.ajax({ url }).then(json => json.features);
-    },
-   	setPlace: function(selected){
-   		this.set('pin', null);
-      var lng = selected.geometry.coordinates[0];
-      var lat = selected.geometry.coordinates[1];
-      var coordinates = [];
-      coordinates.push(lat);
-      coordinates.push(lng);
-      
-  		this.set('place', selected);
-      this.set('pin', coordinates);
-      this.transitionToRoute('index', {queryParams: {pin: this.get('pin'), bbox: null}});
-  	},
-  	clearPlace(){
-  		this.set('place', null);
-  	},
-  	removePin: function(){
-      this.set('pin', null);
-    },
-  	dropPin: function(e){
-      var lat = e.latlng.lat;
-      var lng = e.latlng.lng;
-      var coordinates = [];
-      coordinates.push(lat);
-      coordinates.push(lng);
-      this.set('pin', coordinates);
-    }
+		}
 	}	
 });
