+++ app/stops/controller.js
@@ -1,30 +1,14 @@
 import Ember from 'ember';
 import mapBboxController from 'mobility-playground/mixins/map-bbox-controller';
 import setTextboxClosed from 'mobility-playground/mixins/set-textbox-closed';
+import sharedActions from 'mobility-playground/mixins/shared-actions';
 
-export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
+
+export default Ember.Controller.extend(mapBboxController, setTextboxClosed, sharedActions, {
 	queryParams: ['onestop_id', 'served_by', 'isochrone_mode', 'pin', 'bbox', 'departure_time'],
-  bbox: null,
+
   departure_time: null,
-  leafletBbox: null,
-  leafletBounds: [[37.706911598228466, -122.54287719726562],[37.84259697150785, -122.29568481445312]],
-  pin: null,
   onestop_id: null,
-	pinLocation: Ember.computed('pin', function(){
-    if (typeof(this.get('pin'))==="string"){
-      var pinArray = this.get('pin').split(',');
-      return pinArray;
-    } else {
-      return this.get('pin');
-    }
-  }),
-  place: null,
-  icon: L.icon({
-		iconUrl: 'assets/images/marker1.png',		
-		iconSize: (20, 20),
-    iconAnchor: [10, 24],
-	}),
-	markerUrl: 'assets/images/marker1.png',
   zoom: 12,
 	selectedStop: null,
 	served_by: null,
@@ -35,11 +19,7 @@ export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
 	moment: moment(),
 	mapMoved: false,
 	mousedOver: false,
-  attribution: '&copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors | <a href="http://www.mapzen.com">Mapzen</a> | <a href="http://www.transit.land">Transitland</a> | Imagery © <a href="https://carto.com/">CARTO</a>',
-	closeTextbox: Ember.inject.service(),
-  textboxIsClosed: Ember.computed('closeTextbox.textboxIsClosed', function(){
-    return this.get('closeTextbox').get('textboxIsClosed');
-  }),
+	
   actions: {
 		change(date){
 		},
@@ -76,37 +56,6 @@ export default Ember.Controller.extend(mapBboxController, setTextboxClosed, {
 			this.set('displayIsochrone', false);
 			this.set('isochrones_mode', null);
 		},
-		searchRepo(term) {
-      if (Ember.isBlank(term)) { return []; }
-      const url = `https://search.mapzen.com/v1/autocomplete?api_key=mapzen-jLrDBSP&sources=wof&text=${term}`;
-      return Ember.$.ajax({ url }).then(json => json.features);
-    },
-    setPlace: function(selected){
-      this.set('pin', null);
-      var lng = selected.geometry.coordinates[0];
-      var lat = selected.geometry.coordinates[1];
-      var coordinates = [];
-      coordinates.push(lat);
-      coordinates.push(lng);
-      
-      this.set('place', selected);
-      this.set('pin', coordinates);
-      this.transitionToRoute('index', {queryParams: {bbox: null, pin: this.get('pin')}});
-    },
-  	clearPlace(){
-  		this.set('place', null);
-  	},
-  	removePin: function(){
-      this.set('pin', null);
-    },
-    dropPin: function(e){
-      var lat = e.latlng.lat;
-      var lng = e.latlng.lng;
-      var coordinates = [];
-      coordinates.push(lat);
-      coordinates.push(lng);
-      this.set('pin', coordinates);
-    },
   	setIsochroneMode(mode){
   		if (this.get('isochrone_mode') === mode){
   			this.set('isochrone_mode', null);
